Revert "fix potential linux decoding issue in case coordinates cannot converted to float without precision loss"
This reverts commit 0c0d6c002371f4b16ec2b613f68a7e72b391341d.

+++ Sources/App/Dem/DemController.swift
@@ -4,8 +4,8 @@ import Vapor
 
 struct DemController {
     struct Query: Content {
-        let latitude: [Double]
-        let longitude: [Double]
+        let latitude: [Float]
+        let longitude: [Float]
         
         func validate() throws {
             guard latitude.count == longitude.count else {
@@ -35,7 +35,7 @@ struct DemController {
         let params = try req.query.decode(Query.self)
         try params.validate()
         let elevation = try zip(params.latitude, params.longitude).map { (latitude, longitude) in
-            try Dem90.read(lat: Float(latitude), lon: Float(longitude))
+            try Dem90.read(lat: latitude, lon: longitude)
         }
         return """
         {"elevation":\(elevation)}
